+++ mathics/core/element.py
@@ -419,7 +419,7 @@ class BaseElement(KeyComparable):
         """
         return False
 
-    # Warning - this is going away
+    # Warning: this is going away
     def is_true(self) -> bool:
         """Better use self is SymbolTrue"""
         return False