+++ CHANGES.rst
@@ -59,7 +59,9 @@ Internals
 * Improved support for ``Series`` Issue #46.
 * ``Cylinder`` rendering is implemented in Asymptote.
 * ``N[_,_,Method->method]`` was reworked (Issue #137).
-
+* ``InstanceableBuiltin`` -> ``BuiltinElement``
+* ``BoxConstruction`` -> ``BoxExpression``
+ 
 
 Package update
 ++++++++++++++